# internal/web/swagger/openapi.yaml
openapi: 3.1.0
info:
  title: harbrr management API
  version: 0.1.0
  description: >
    harbrr's HTTP surface: the management API (first-run setup, authentication,
    API-key management, indexer and app-sync configuration) plus the entry points of
    the *arr-facing Torznab feed. The Torznab wire contract itself — the XML body
    shapes — is defined by Jackett/Newznab and served by internal/torznab; here we
    document only its HTTP routes, parameters, and content types (see
    docs/architecture.md). Hand-authored, embedded via //go:embed, and guarded by a
    drift test that asserts the spec matches the mounted routes (make test-openapi).

    Authentication: the management API uses a cookie session (after login) or the
    X-API-Key header; the Torznab feed authenticates with an apikey (or passkey)
    query parameter, as *arr clients expect. CSRF: a cookie-authenticated mutating
    request (POST/PUT/PATCH/DELETE) must echo the session's CSRF token in an
    X-CSRF-Token header — the token is minted at login and handed to the client via
    the non-HttpOnly harbrr_csrf cookie and GET /api/auth/me; a missing or wrong
    token is 403. X-API-Key and auth-disabled callers are exempt (see
    docs/security.md).
  license:
    name: GPL-2.0-or-later
    url: https://www.gnu.org/licenses/gpl-2.0.html
...
paths:
...
  /api/app-connections/{id}/test:
    post:
      summary: Test a connection's reachability and credentials
      operationId: testConnection
      tags: [App Connections]
      security: [{ session: [] }, { apiKey: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        "200":
          description: the test result (ok=false carries a scrubbed error)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TestResult" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/download-clients:
    get:
      summary: List download clients
      operationId: listDownloadClients
      tags: [Download Clients]
      security: [{ session: [] }, { apiKey: [] }]
      responses:
        "200":
          description: the download clients (secrets redacted)
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/DownloadClient" } }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      summary: Add a download client
      description: Stores the client's secret encrypted at rest; it reads back redacted.
      operationId: createDownloadClient
      tags: [Download Clients]
      security: [{ session: [] }, { apiKey: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateDownloadClient" }
      responses:
        "201":
          description: the created download client
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DownloadClient" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /api/download-clients/{id}:
    get:
      summary: Get one download client
      operationId: getDownloadClient
      tags: [Download Clients]
      security: [{ session: [] }, { apiKey: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        "200":
          description: the download client
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DownloadClient" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      summary: Update a download client (a new secret rotates the credential; kind is immutable)
      operationId: updateDownloadClient
      tags: [Download Clients]
      security: [{ session: [] }, { apiKey: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateDownloadClient" }
      responses:
        "204": { description: updated }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      summary: Delete a download client
      operationId: deleteDownloadClient
      tags: [Download Clients]
      security: [{ session: [] }, { apiKey: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        "204": { description: deleted }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/download-clients/{id}/test:
    post:
      summary: Test a download client's connection
      description: >
        Confirms the configured client is reachable with its stored credentials. A
        pass is {"ok":true}; a connection failure is 200 with ok=false and a
        scrubbed error; an unknown id is 404.
      operationId: testDownloadClient
      tags: [Download Clients]
      security: [{ session: [] }, { apiKey: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        "200":
          description: the test outcome
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TestResult" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/download-clients/{id}/enable:
    post:
      summary: Enable a download client
      operationId: enableDownloadClient
      tags: [Download Clients]
      security: [{ session: [] }, { apiKey: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        "204": { description: enabled }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/download-clients/{id}/disable:
    post:
      summary: Disable a download client
      operationId: disableDownloadClient
      tags: [Download Clients]
      security: [{ session: [] }, { apiKey: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        "204": { description: disabled }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: harbrr_session
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: >
        A bearer credential — send over HTTPS only; it must never traverse a
        cleartext link. Terminate TLS at harbrr or a trusted reverse proxy.
  responses:
    BadRequest:
      description: invalid request
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: authentication required or failed
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    NotFound:
      description: resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string, description: human-readable message }
        code:
          type: string
          description: >
            machine-readable code clients can branch on — one of: bad_request,
            invalid, unauthorized, invalid_credentials, invalid_api_key, forbidden,
            not_found, conflict, already_setup, not_implemented, internal. An OIDC
            ID-token verification failure also uses invalid_credentials.
      required: [error, code]
...
    TestResult:
      type: object
      description: >
        The outcome of a test action. ok=false carries a credential-scrubbed error
        message; the HTTP status is 200 even on a credential/connectivity failure.
      properties:
        ok: { type: boolean }
        error: { type: string }
      required: [ok]
...
    QBittorrentSettings:
      type: object
      description: qBittorrent-specific per-client options; all fields are optional.
      properties:
        category: { type: string }
        tags: { type: array, items: { type: string } }
        startPaused: { type: boolean }
        tlsSkipVerify: { type: boolean }
    QuiSettings:
      type: object
      description: >
        qui-specific per-client options. qui (autobrr/qui) is a multi-instance
        qBittorrent manager keyed by int instance id — instanceId is required.
      properties:
        instanceId: { type: integer, description: "required; > 0" }
        category: { type: string }
        tags: { type: array, items: { type: string } }
        startPaused: { type: boolean }
      required: [instanceId]
    FloodSettings:
      type: object
      description: >
        Flood-specific per-client options. Flood has no category concept — a
        caller's category is folded into tags on add.
      properties:
        destination: { type: string }
        tags: { type: array, items: { type: string } }
        startPaused: { type: boolean }
    DownloadStationSettings:
      type: object
      description: Synology Download Station-specific per-client options.
      properties:
        directory: { type: string, description: "relative to a shared folder; no leading slash" }
    BlackholeSettings:
      type: object
      description: >
        blackhole watch-folder configuration. At least one of torrentDir/nzbDir must
        be set; a payload whose protocol has no configured dir fails as unsupported.
      properties:
        torrentDir: { type: string, description: "absolute path a .torrent/.magnet is written into" }
        nzbDir: { type: string, description: "absolute path a .nzb is written into" }
        saveMagnetFiles: { type: boolean, description: "write a magnet-only release as a .magnet file instead of failing" }
    DownloadClientSettings:
      type: object
      description: >
        Kind-specific options, keyed by kind — at most the one field matching the
        client's own kind is populated.
      properties:
        qbittorrent: { $ref: "#/components/schemas/QBittorrentSettings" }
        blackhole: { $ref: "#/components/schemas/BlackholeSettings" }
        qui: { $ref: "#/components/schemas/QuiSettings" }
        flood: { $ref: "#/components/schemas/FloodSettings" }
        downloadStation: { $ref: "#/components/schemas/DownloadStationSettings" }
    DownloadClient:
      type: object
      description: >
        A configured download client harbrr can send grabbed releases to.
        host/username are plain; the secret (password or API key, depending on
        kind) is write-only and is never echoed back — it reads back as the
        <redacted> sentinel.
      properties:
        id: { type: integer, format: int64 }
        name: { type: string }
        kind:
          type: string
          enum: [qbittorrent, blackhole, qui, flood, download-station]
        appId: { type: integer, format: int64, nullable: true, description: the App holding this client's identity + credential; null for host-less kinds (blackhole) (ADR 0004) }
        enabled: { type: boolean }
        host: { type: string, description: enriched from the App (blank for host-less kinds) }
        username: { type: string }
        secret: { type: string, description: "the client's secret, redacted (<redacted>)" }
        settings: { $ref: "#/components/schemas/DownloadClientSettings" }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, name, kind, enabled, host, username, secret, settings, createdAt, updatedAt]
    CreateDownloadClient:
      type: object
      description: >
        A networked kind references the App by appId (reuse) OR inline (host [+ username
        + secret], get-or-create). A host-less kind (blackhole) takes neither — Settings
        only.
      properties:
        name: { type: string }
        kind:
          type: string
          enum: [qbittorrent, blackhole, qui, flood, download-station]
        appId: { type: integer, format: int64, description: reuse an existing App (omit host/username/secret) }
        host: { type: string, description: "inline App identity (get-or-create); must be empty for host-less kinds (blackhole)" }
        username: { type: string }
        secret: { type: string, description: "optional; sealed on the App" }
        settings: { $ref: "#/components/schemas/DownloadClientSettings" }
      required: [name, kind]
    UpdateDownloadClient:
      type: object
      description: >
        A partial update of the client's surface fields; omitted fields are unchanged.
        Identity + credential (host, username, secret) are App-level now. Kind is immutable.
      properties:
        name: { type: string }
        settings: { $ref: "#/components/schemas/DownloadClientSettings" }
...